feat(Login): add useAuth
