+++ src/components/Header.vue
@@ -11,7 +11,8 @@
             <li><a href="#search">Search</a></li> 
         </ul>
         <div class="cc">
-           <RouterLink to="login" class="signInbtn">Sign In</RouterLink>
+           <RouterLink v-if="!user.id" to="login" class="signInbtn">Sign In</RouterLink>
+           <h3 v-else >Profile</h3>
            <div class="icon" @click="openMenu">
               <span class="fas fa-bars"></span>
             </div>
@@ -20,7 +21,8 @@
         <div class="menuBurger" ref="menuBurgerRef">
             <ul>
                 <li><a class="fas fa-times close-hamburger"></a></li>
-                <li><a class="">Sign In</a></li>
+                <li v-if="!user.id"><RouterLink to="login">Sign In</RouterLink></li>
+                <li v-else ><h3>Profile</h3></li>
                 <li><a href="#home">Home</a></li>
                 <li><a href="#movies">Series</a></li>
                 <li><a href="#coming">Movie</a></li>
@@ -31,9 +33,13 @@
 </template>
 
 <script setup>
-import { ref } from 'vue';
+import { inject, ref } from 'vue';
 import { RouterLink } from 'vue-router';
+const user = inject('user');
+console.log(user);
 const menuBurgerRef = ref(null);
+
+
 
 
 const openMenu = () =>{
